backend/requirements.txt
Flask
Flask-Cors
PyJWT
passlib[bcrypt]
python-dotenv
